refactor(types): reorganize and consolidate McpServer and LLMConfig types into separate files for improved structure and clarity
