+++ sites/playground/server/src/types/index.ts
@@ -1,1 +1,3 @@
 export * from './playground-config.js';
+export * from './mcp-server.js';
+export * from './llm-config.js';
