+++ sites/playground/server/src/types/playground-config.ts
@@ -1,5 +1,5 @@
 import type { PlaygroundAgentConfig } from '../a2a-tools/index.js';
-import type { McpServersConfig } from '../chat-genui.js';
+import type { McpServersConfig } from './mcp-server.js';
 
 export interface IPlaygroundConfig {
   mcpServers: McpServersConfig;
@@ -8,4 +8,4 @@ export interface IPlaygroundConfig {
   model: string;
   temperature: number;
   agents: PlaygroundAgentConfig[];
-};
+}
